refactor: Move to not depend on a domain QUZ-53
