+++ game-service/game-infra/src/main/kotlin/com/grepp/quizy/game/infra/repository/GameRedisRepository.kt
@@ -1,6 +1,6 @@
 package com.grepp.quizy.game.infra.repository
 
-import com.grepp.quizy.game.infra.redis.entity.GameRedisEntity
+import com.grepp.quizy.game.infra.game.GameRedisEntity
 import org.springframework.data.repository.CrudRepository
 
 interface GameRedisRepository : CrudRepository<GameRedisEntity, Long> {
